scans-analyzer/requirements.txt
flask
flask-cors
opencv-python
numpy
Pillow
transformers
torch
torchvision
pytesseract
requests